logos_system/in_development/aug01_validator_hub/requirements.txt
PyYAML>=5.4
pytest>=7.0

pgmpy>=0.2.8

statsmodels>=0.13
pmdarima>=1.8.4

PyWavelets>=1.3.0

sentence-transformers>=2.2.2

torch>=1.13.1
torchvision>=0.14.1

networkx>=2.8
